Changes on the CMS.
There is a new Javascript class and the web module is derived
into many files. ipfire.py behaves different and there is
a new info.json file that should make some things easier.

I also hope that it will be faster or even as fast as it has
been before these changes.